requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0
drf-yasg>=1.21,<2.0
psycopg2-binary>=2.9,<3.0
pytest-django>=4.5,<5.0
django-filter>=23.1,<24.0
requests==2.32.3